BMP refactor: move BMPInput and BMPOutput to the files where they are used.
